Unit test

+++ league-service/src/main/java/com/works/leagueservice/domain/Constants.java
@@ -6,6 +6,6 @@ package com.works.leagueservice.domain;
  */
 public class Constants {
 
-    public static final int DEFAULT_INVALID_VALUE = 0;
-    public static final int DEFAULT_VALID_VALUE = 1;
+    public static final Long DEFAULT_INVALID_VALUE = 0L;
+    public static final Long DEFAULT_VALID_VALUE = 1L;
 }
